v2.0.4
EN:

Experience points are no longer awarded during warm-up sessions.
Fixed errors that were found in the console; however, there is no full confidence that all errors have been corrected.
A new event has been added, for which experience is awarded: bomb planting.
A new event has been added, for which experience is awarded: bomb detonation.
A new event has been added, for which experience is awarded: picking up the bomb.
A new event has been added, for which experience is awarded: throwing the bomb.
A new event has been added, for which experience is awarded: shooting through.
A new configuration file, Weapons.yml, has been added, in which you can set the amount of experience awarded for kills with different types of weapons, for example, 20 points of experience for a kill with a knife.
IMPORTANT: It is necessary to recreate the Config.yml configuration file for the plugin to work correctly.

RU:

Во время разминки опыт больше не начисляется.
Исправлены ошибки, обнаруженные в консоли; однако, нет полной уверенности, что все ошибки исправлены.
Добавлено новое событие, за выполнение которого начисляется опыт: установка бомбы.
Добавлено новое событие, за выполнение которого начисляется опыт: взрыв бомбы.
Добавлено новое событие, за выполнение которого начисляется опыт: поднятие бомбы.
Добавлено новое событие, за выполнение которого начисляется опыт: выброс бомбы.
Добавлено новое событие, за выполнение которого начисляется опыт: прострел.
Добавлен новый конфигурационный файл Weapons.yml, в котором можно настроить количество опыта, начисляемого за убийство с использованием различных видов оружия, например, за убийство ножом начисляется 20 очков опыта.
ВАЖНО: Необходимо пересоздать конфигурационный файл Config.yml для корректной работы плагина.