v2.0.5
EN:

Presumably fixed server crashes.
Now can disable your clan tag by entering the command !tagrank
The ability to disable clan tags for all users is now available. To do this, in the configuration file, in EnableClanTags, set the value to False.
Added a multiplier for increased experience for using a website link in the nickname. The multiplier and link can be configured in the config, and there is also an option to disable this feature.
Fixed errors where actions like tossing, picking up, or planting the bomb were awarded experience, even if the necessary number of players was not present.
Added new events for which experience is now also awarded: picking up a hostage, delivering him to the player's base, and deductions for losing a hostage.
IMPORTANT: To activate the new parameters, you need to recreate the main configuration file Config.yml. I might have forgotten to mention some changes, as a lot of work was done.
RU:

Предположительно устранены сбои.
Теперь можно отключить свой клан-тег, введя команду !tagrank
Возможность отключения клан-тегов для всех пользователей теперь доступна. Для этого в конфигурационном файле, в EnableClanTags, необходимо установить значение False.
Добавлен коэффициент увеличения опыта за использование в никнейме ссылки на сайт. Множитель и ссылку можно настроить в конфиге, также там предусмотрена возможность отключения этой функции.
Исправлены ошибки, когда за действия, такие как подбрасывание, поднятие или установка бомбы, начислялся опыт, даже если не было необходимого количества игроков.
Добавлены новые события, за которые теперь также начисляется опыт: поднятие заложника, его доставка на базу игрока и снятие за потерю заложника.
ВАЖНО: Для активации новых параметров необходимо пересоздать основной конфигурационный файл Config.yml. Возможно, я забыл упомянуть о некоторых изменениях, так как было проделано много работы.